docs: enforce 90 percent coverage on changes
